% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_cod_name}
\alias{.add_cod_name}
\title{Add code and station name from stations metadata
Add code and station name from stations metadata}
\usage{
.add_cod_name(DT, txt_file)
}
\arguments{
\item{txt_file}{path to ASCII file}

\item{x}{tibble}
}
\value{
original data with extra columns \code{code_stn} and \code{name_stn}.
}
\description{
Add code and station name from stations metadata
Add code and station name from stations metadata
}
\details{
It is used in \code{import_qnat}
}
\seealso{
"\code{\link[import_metadata]{HEobs}}"
}
