% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.informative}
\alias{.informative}
\title{Drops variables in DT which have the same values for all observations.}
\usage{
.informative(DT)
}
\arguments{
\item{DT}{\code{data.frame}, \code{tibble} or \code{data.table}.}
}
\value{
\code{\link[tibble]{as_tibble}} of original data excluding
non-informative variables.
}
\description{
Drops variables in DT which have the same values for all observations.
}
\details{
It is used in \code{import_metadata}.
}
\seealso{
\code{\link[data.table]{as.data.table}}, \code{\link[tibble]{as_tibble}},
"\code{\link[import_metadata]{HEobs}}"
}
