% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Import metadata from streamflow stations of ONS}
\usage{
extract_metadata(txt_file = "", informative = FALSE)
}
\arguments{
\item{txt_file}{Character. File path (or link to data file).}

\item{informative}{Logical. Drop non-informative variables. Default: FALSE}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Import metadata from streamflow stations of ONS
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  file <- "https://www.dropbox.com/s/d40adhw66uwueet/VazoesNaturaisONS_D_87UHEsDirceuAssis_2018.dat?dl=1"
  qnat_meta <- import_metadata(file, informative = TRUE)
  str(qnat_meta)
 }
}
}
\seealso{
\code{\link[rio]{import}}
\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{slice}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{across}}
\code{\link[tibble]{as_tibble}}
\code{\link[tidyselect]{vars_select_helpers}}
\code{\link[janitor]{clean_names}}
}
