% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HEobs-package.R
\docType{package}
\name{HEobs-package}
\alias{HEobs-package}
\title{\pkg{HEobs}: Data Processing Of Natural Streamflow And Metadata From Ons Stations}
\description{
The goal of \pkg{HEobs} is to Ingest, clean and tidy data of natural
streamflow and metadata from ONS stations.
}
\section{HEobs functions}{


\itemize{
\code{\link{import_qnat}}: Import data file of daily naturalized streamflow

\code{\link{extract_metadata}}: Import metadata from streamflow stations of ONS

\code{\link{wider}}: Pivot data from long to wide format.
}
}

