% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnat_data.R
\name{import_qnat}
\alias{import_qnat}
\title{Import data file of daily naturalized streamflow}
\usage{
import_qnat(file, complete = TRUE, add_stn_info = TRUE)
}
\arguments{
\item{file}{Character. Path to ascii data file (or a URL).}

\item{complete}{logical, Default: TRUE. Make missing dates explicit using
\code{complete_dates()}.}

\item{add_stn_info}{logical, Default: TRUE. Get code and station name
from stations metadata.}
}
\value{
a \link[tibble:tibble-package]{tibble} with tidy data.
}
\description{
Import data file of daily naturalized streamflow
}
\details{
The source ascii file contains data and metadata from ONS Hydroelectric
Plants. Metadata is extracted with \link{extract_metadata} while the data is
extract here.
}
\examples{
if (TRUE) {
  qnat <- import_qnat(NA_character_, complete = TRUE, add_stn_info = TRUE)
  str(qnat)
}
}
