% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnat_data.R
\name{wider}
\alias{wider}
\title{Pivot data from long to wide format}
\usage{
wider(
  qnat,
  names_from = "code_stn",
  names_prefix = "stn_",
  values_from = "qnat"
)
}
\arguments{
\item{qnat}{a \link[tibble:tibble-package]{tibble} with tidy data. Usually the
output from \link{import_qnat}.}

\item{names_from}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> A pair of
arguments describing which column (or columns) to get the name of the
output column (\code{names_from}), and which column (or columns) to get the
cell values from (\code{values_from}).

If \code{values_from} contains multiple values, the value will be added to the
front of the output column.}

\item{names_prefix}{String added to the start of every variable name. This is
particularly useful if \code{names_from} is a numeric vector and you want to
create syntactic variable names.}

\item{values_from}{<\code{\link[tidyr:tidyr_tidy_select]{tidy-select}}> A pair of
arguments describing which column (or columns) to get the name of the
output column (\code{names_from}), and which column (or columns) to get the
cell values from (\code{values_from}).

If \code{values_from} contains multiple values, the value will be added to the
front of the output column.}
}
\value{
a widen \link[tibble:tibble-package]{tibble} version from
input data (\code{qnat}). Each column will correspond to a time series of a
station. Default for variable names will be a string like \verb{qnat_\{names_from\}}
. Most of the time the value of names_from is \code{code_stn} or \code{id}.
}
\description{
Pivot data from long to wide format
}
\details{
This almost a wrapper function of \code{tidyr::pivot_wider()}.
}
\examples{
if(FALSE){
 qnat <- import_qnat(NA_character_, complete = TRUE, add_stn_info = TRUE)
 str(qnat)
 qnat_wide <- wider(qnat)
 str(qnat_wide)
}
}
