% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Import metadata from streamflow stations of ONS}
\source{
The metadata 87 hydroelectric power plants operated by ONS were
supplied by Saul Aires (ANA skilled in Water Resources) by Prof. Carlos
Lima (UnB), by email on 2020-03-17.
}
\usage{
extract_metadata(file, informative = FALSE)
}
\arguments{
\item{file}{Character. Path to ascii data file (or a URL).}

\item{informative}{Logical. Drop non-informative variables. Default: FALSE}
}
\value{
a \link[tibble:tibble-package]{tibble} with tidy data
}
\description{
Import metadata from streamflow stations of ONS
}
\details{
The text file contains data and metadata from ONS Hydroelectric
Plants. Metadata is extracted from the first 15 lines of the ascii file.
The information for each HPP is stored in pairs of columns. The first column
is the name of the variables and the second the values of the variables.
The data is processed in tidy format.
}
\examples{
if (FALSE) {
  file <- data_link
  qnat_meta <- extract_metadata(file, informative = TRUE)
  str(qnat_meta)
}
}
